+++ backend/src/main/java/com/yagubogu/auth/service/JwtProvider.java
@@ -1,4 +1,4 @@
-package com.yagubogu.auth.token;
+package com.yagubogu.auth.service;
 
 
 import com.auth0.jwt.JWT;
