+++ backend/src/test/java/com/yagubogu/auth/token/JwtProviderTest.java
@@ -6,6 +6,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import com.yagubogu.auth.dto.MemberClaims;
 import com.yagubogu.auth.config.JwtProperties;
 import com.yagubogu.auth.config.JwtProperties.TokenProperties;
+import com.yagubogu.auth.service.JwtProvider;
 import com.yagubogu.global.exception.UnAuthorizedException;
 import com.yagubogu.member.domain.Role;
 import org.junit.jupiter.api.BeforeEach;
